update build script

+++ build.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
-
-IMAGE='jdimpson/yt-contained';
+set -e
+IMAGE=$(cat IMAGENAME);
+if test -z "$IMAGE"; then
+	echo "Problem reading image name from file IMAGENAME";
+	exit 1;
+fi
 
 if docker images | grep -q "$IMAGE"; then
 	docker image rm "$IMAGE";
